implemented a parser for console commands as well as tests for it
For a given text command (String) the parser finds the corresponding
element of the Command enum. Additionally, for a correctly entered
text command it can extract the arguments, e.g. for:
"contacts add Alice 127.0.0.1 4444" it would be able to match the String
to the command Command.CONTACTS_ADD and would extract the arguments
["Alice", "127.0.0.1", "4444"]