Added Basic Network Test
The BasicNetworkTest.jar can be used to test if two PCs can communicate with each other. Both parties run it (on Win with >java -jar BasicNetworkTest.jar) and follow the on-screen instructions.
One party should play the role of the server, the other the role of the client.
Once connected, the client can send a message to the server, which in turn can respond to that message. This the loops until one party closes the console or looses connection.