WIP please read extended message, does not compile atm
sending and receiving of files is more or less implemented, however, it
requires that bytes can be sent via connections, i made some changes so
that this is possible, however, a lot of classes still work with Strings
in many places including, for example, ConnectionEndpoint, which uses
them internally. I could try to rework all of that, however, Jonas
reworked the entire Connection system in the gui branch regardless, so I
would just be wasting time.

Instead, what is most likely needed to properly compile and test this is
a rebase once Jonas PR is through, or maybe a new branch where we just
copy and paste things over if a rebase is a giant headache, I don't
know.