removed deleteCurrent parameter from generateSignatureKeys
The parameter was confusing and not needed - if the current keys are
desired to be deleted, the method for that purpose can be used.
Personally, I think the only parameters the method should have are the
name of the keys to generate and overwrite, to adhere to the "single
responsibility principle" as much as possible. But setAsKeyFile is still
reasonable, so I think it is fine to keep.