initial implementation of key index synchronization
still WIP but the general idea is:
- A wants to send encrypted file to B
- A and B have one and the same key
- A thinks they are currently at index n, but does not know which index
B thinks they are at
- A tells B "I want to encrypt starting at Index n"
- B receives this message and looks if index n is >= index m, with index
m being the index B is currently at. If it is equal or larger, that
means A will not be using any bytes that have already been used.
-- n >= m
--- B tells A "ok you can do the encryption"
-- n < m
--- B tells A "no you can not do encryption, we are at index m already"
- A has queued up the encrypted message c to send
- now A gets a message from B that is the answer to the request it sent
-- if answer is yes, A goes ahead and sends c
-- if answer is no, A increases their index n to m and does not send c

currently works for text messages in good circumstances, files and edge
cases are not tested yet but will come soon (tm)